+++ src/app/code/community/EbayEnterprise/PayPal/Block/Express/Form.php
@@ -31,7 +31,7 @@ class EbayEnterprise_PayPal_Block_Express_Form extends Mage_Payment_Block_Form
         $markClass = Mage::getConfig()->getBlockClassName(static::PAYMENT_MARK);
         $mark = new $markClass();
         $markHtml = $mark->toHtml();
-        $helper = Mage::helper('paypal');
+        $helper = Mage::helper('ebayenterprise_paypal');
         $translatedRedirectMessage = $helper->__(static::WILL_REDIRECT_MESSAGE);
         $this->setMethodTitle(''); // Title conflicts with PayPal mark
         $this->setMethodLabelAfterHtml($markHtml);